+++ src/api/views/websites.py
@@ -2,21 +2,32 @@
 # Standard Python Libraries
 from datetime import datetime
 import io
+import logging
+import os
 import shutil
 from uuid import uuid4
 
 # Third-Party Libraries
 import boto3
-from flask import jsonify, request, send_file
+from flask import current_app, jsonify, request, send_file
 from flask.views import MethodView
 import requests
+from selenium import webdriver
 
 # cisagov Libraries
-from api.manager import ApplicationManager, WebsiteManager
+from api.manager import (
+    ApplicationManager,
+    CategoryManager,
+    ProxyManager,
+    WebsiteManager,
+)
 from settings import STATIC_GEN_URL, WEBSITE_BUCKET
 from utils.aws.redirect_handler import delete_redirect, modify_redirect, setup_redirect
 from utils.aws.site_handler import delete_site, launch_site
+from utils.two_captcha import two_captcha_api_key
 
+category_manager = CategoryManager()
+proxy_manager = ProxyManager()
 website_manager = WebsiteManager()
 application_manager = ApplicationManager()
 route53 = boto3.client("route53")
@@ -52,6 +63,25 @@ class WebsiteView(MethodView):
         """Get Website details."""
         website = website_manager.get(document_id=website_id)
         return jsonify(website)
+
+    def put(self, website_id):
+        """Update website."""
+        website = website_manager.get(document_id=website_id)
+        if request.json.get("application"):
+            application = application_manager.get(
+                filter_data={"name": request.json["application"]}
+            )
+            website["application_id"] = application["_id"]
+            # Save application to history
+            website["history"] = website.get("history", [])
+            website["history"].append(
+                {
+                    "application": application,
+                    "launch_date": datetime.utcnow(),
+                }
+            )
+
+        return jsonify(website_manager.update(document_id=website_id, data=website))
 
     def delete(self, website_id):
         """Delete website and hosted zone."""
@@ -90,25 +120,6 @@ class WebsiteContentView(MethodView):
             attachment_filename=f"{website['name']}.zip",
             mimetype="application/zip",
         )
-
-    def put(self, website_id):
-        """Update website."""
-        website = website_manager.get(document_id=website_id)
-        if request.json.get("application"):
-            application = application_manager.get(
-                filter_data={"name": request.json["application"]}
-            )
-            website["application_id"] = application["_id"]
-            # Save application to history
-            website["history"] = website.get("history", [])
-            website["history"].append(
-                {
-                    "application": application,
-                    "launch_date": datetime.utcnow(),
-                }
-            )
-
-        return jsonify(website_manager.update(document_id=website_id, data=website))
 
     def post(self, website_id):
         """Upload files and serve s3 site."""
@@ -302,3 +313,79 @@ class WebsiteRecordView(MethodView):
         )["route53"]["id"]
         resp = route53.list_resource_record_sets(HostedZoneId=hosted_zone_id)
         return jsonify(resp["ResourceRecordSets"])
+
+
+class WebsiteCategorizeView(MethodView):
+    """WebsiteCategorizeView."""
+
+    def get(self, website_id):
+        """Manage categorization of active sites."""
+        browserless_endpoint = os.environ.get("BROWSERLESS_ENDPOINT")
+        chrome_options = webdriver.ChromeOptions()
+        chrome_options.add_argument("--no-sandbox")
+        chrome_options.add_argument("--headless")
+        active_site = website_manager.get(document_id=website_id)
+        domain = active_site.get("domain").get("Name")
+        domain_url = domain[:-1]
+        if active_site.get("is_categorized"):
+            return {"error": f"{domain} has already been categorized."}
+
+        category = category_manager.get(
+            filter_data={"name": request.args.get("category")}
+        )
+
+        if not category:
+            return {"error": "Category does not exist"}
+
+        is_submitted = []
+        # Submit domain to proxy
+        if not current_app.config["TESTING"]:
+            proxies = proxy_manager.all()
+            for proxy in proxies:
+                proxy_name = proxy["name"]
+
+                # Get unique category name for each proxy
+                proxy_category = "".join(
+                    detail.get(proxy_name)
+                    for detail in category.get("proxies")
+                    if proxy_name in detail
+                )
+
+                try:
+                    driver = webdriver.Remote(
+                        command_executor=f"http://{browserless_endpoint}/webdriver",
+                        desired_capabilities=chrome_options.to_capabilities(),
+                    )
+                    driver.set_page_load_timeout(60)
+                    exec(
+                        proxy.get("script").decode(),
+                        {
+                            "driver": driver,
+                            "url": proxy.get("url"),
+                            "domain": domain_url,
+                            "api_key": two_captcha_api_key,
+                            "category": proxy_category,
+                        },
+                    )
+                    driver.quit()
+                    is_submitted.append(
+                        {
+                            "_id": proxy["_id"],
+                            "name": proxy_name,
+                            "is_categorized": False,
+                        }
+                    )
+                    logging.info(f"Categorized with {proxy_name}")
+                except Exception as err:
+                    driver.quit()
+                    logging.error(f"{proxy_name} has failed")
+                    return {"error": str(err)}
+
+        # Quit WebDriver
+        driver.quit()
+
+        # Update database
+        website_manager.update(
+            document_id=website_id, data={"is_submitted": is_submitted}
+        )
+        return jsonify({"message": f"{domain} has been successfully categorized"})
